requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
scipy>=1.11
pymongo[srv]>=4.5
python-dotenv>=1.0
pyyaml>=6.0
mlflow>=2.10
joblib>=1.3
matplotlib>=3.7
seaborn>=0.13
pyarrow>=15.0
lightgbm>=4.0
pytest>=7.4
pytest-cov>=4.1
streamlit>=1.44
